# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API Documentation",
        "description": "Swagger Docs for Dashboard APIs (Admin Panel)",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/dashboard/products": {
            "get": {
                "tags": [
                    "Dashboard - Product"
                ],
                "summary": "Get all products with category and pagination",
                "operationId": "6481e98f63350a26907aace97aade7ae",
                "responses": {
                    "200": {
                        "description": "All products retrieved with pagination"
                    },
                    "400": {
                        "description": "No products found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dashboard - Product"
                ],
                "summary": "Create a new product",
                "operationId": "ae6c45e253cdbe3f31e9d3a3a8eaf6cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "price",
                                    "description",
                                    "category_id",
                                    "image"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "iPhone 14"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 999.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Latest Apple phone"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation failed or image missing"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/products/{id}": {
            "get": {
                "tags": [
                    "Dashboard - Product"
                ],
                "summary": "Get a single product by ID",
                "operationId": "36e509e84ba147431a94b601d521b692",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dashboard - Product"
                ],
                "summary": "Update an existing product",
                "operationId": "14670ca7abfab96c36775fc87ff43e25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "iPhone 15"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1099.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated phone"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dashboard - Product"
                ],
                "summary": "Delete a product by ID",
                "operationId": "a71183f1e44533721d599a207081ecf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}